feat: complete catalog sync infrastructure and resolve activity log visibility
🎉 MULTI-CATALOG SYSTEM IMPLEMENTATION COMPLETED

Fixed Critical Issue:
- Resolved "Bicycle Crunches" not appearing in activity log table
- Root cause: Missing catalog_id field in activity logs sync infrastructure

Database & Sync Infrastructure:
- Added catalog_id column to activity_logs table via migration
- Updated sync_v2 edge function (v36) with complete catalog support:
  * exercise_catalogs added to SYNC_TABLES
  * catalog_id added to exercises and activity_logs MUTABLE_FIELD_ALLOWLIST
  * Complete exercise_catalogs field allowlist implemented
- Updated IndexedDB schema to version 19 with catalog_id support
- Applied migrations to both dev and production environments

Application Logic:
- Fixed all 4 activity log creation instances in App.tsx to include catalog_id
- Updated ActivityLog interface to include catalog_id field
- Enhanced sync infrastructure to handle catalog relationships

Testing & Validation:
- Added comprehensive catalog_id database schema tests
- Verified complete sync infrastructure deployment
- Validated TypeScript compilation with zero errors
- Confirmed backward compatibility with existing functionality

Documentation:
- Updated implementation plan with complete progress tracking
- Comprehensive CHANGELOG.md with technical details
- All 4 phases of multi-catalog system marked as completed
